For silly reasons, this removes a matplotlib warning.

+++ tutorial/step_3.py
@@ -195,7 +195,7 @@ if __name__=='__main__':
     right_edge = max(vs_xedge[-1], vphi_xedge[-1], rho_xedge[-1])
     bottom_edge= 4.3
     top_edge=11.3
-    aspect_ratio = (right_edge-left_edge)/(top_edge-bottom_edge)
+    aspect_ratio = float((right_edge-left_edge)/(top_edge-bottom_edge))
     gamma = 0.5  #Mess with this to change intensity of colormaps near the edges
 
     plt.subplot(111, aspect='equal')
